+++ demos/src/Experiments/CollaborationAnnotation/Vue/extension/AnnotationState.ts
@@ -1,6 +1,10 @@
 import { EditorState, Transaction } from '@tiptap/pm/state'
 import { Decoration, DecorationSet } from '@tiptap/pm/view'
-import { absolutePositionToRelativePosition, relativePositionToAbsolutePosition, ySyncPluginKey } from 'y-prosemirror'
+import {
+  absolutePositionToRelativePosition,
+  relativePositionToAbsolutePosition,
+  ySyncPluginKey,
+} from '@tiptap/y-tiptap'
 import * as Y from 'yjs'
 
 import { AnnotationItem } from './AnnotationItem.js'