+++ demos/src/Extensions/Selection/React/index.spec.js
@@ -1,0 +1,9 @@
+context('/src/Extensions/Selection/React/', () => {
+  beforeEach(() => {
+    cy.visit('/src/Extensions/Selection/React/')
+  })
+
+  it('should have class', () => {
+    cy.get('.tiptap span:first').should('have.class', 'selection')
+  })
+})