+++ packages/core/src/index.ts
@@ -21,10 +21,4 @@ export * from './utilities/index.js'
 export interface Commands<ReturnType = any> {}
 
 // eslint-disable-next-line
-export interface ExtensionConfig<Options = any, Storage = any> {}
-
-// eslint-disable-next-line
-export interface NodeConfig<Options = any, Storage = any> {}
-
-// eslint-disable-next-line
-export interface MarkConfig<Options = any, Storage = any> {}
+export interface Storage {}