+++ packages/extension-collaboration-cursor/src/collaboration-cursor.ts
@@ -1,6 +1,6 @@
 import { Extension } from '@tiptap/core'
 import { DecorationAttrs } from '@tiptap/pm/view'
-import { defaultSelectionBuilder, yCursorPlugin } from 'y-prosemirror'
+import { defaultSelectionBuilder, yCursorPlugin } from '@tiptap/y-tiptap'
 
 type CollaborationCursorStorage = {
   users: { clientId: number; [key: string]: any }[]
@@ -75,6 +75,10 @@ declare module '@tiptap/core' {
        */
       user: (attributes: Record<string, any>) => ReturnType
     }
+  }
+
+  interface Storage {
+    collaborationCursor: CollaborationCursorStorage
   }
 }