+++ tests/cypress/integration/core/onContentError.spec.ts
@@ -150,12 +150,12 @@ describe('onContentError', () => {
       enableContentCheck: true,
       onContentError: args => {
         args.disableCollaboration()
-        expect(args.editor.storage.collaboration.isDisabled).to.eq(true)
+        expect(args.editor.storage.collaboration).to.eq(undefined)
       },
     })
 
     expect(editor.getText()).to.eq('')
-    expect(editor.storage.collaboration.isDisabled).to.eq(true)
+    expect(editor.storage.collaboration).to.eq(undefined)
   })
 
   it('does not remove the collaboration extension when has valid content (when enableContentCheck = true)', () => {